requirements.txt
python-telegram-bot==20.8
flask
flask-cors
python-dotenv
